+++ src/config/nflGeography.ts
@@ -1,5 +1,0 @@
-const config = {
-  spreadsheetId: process.env.NFL_GEOGRAPHY_SPREADSHEET_ID,
-};
-
-export default config;