+++ src/projects/NFLGeoguessr.tsx
@@ -1,0 +1,15 @@
+export const NFLGeoguessr = {
+  name: 'NFL Geoguessr',
+  color: {
+    border: {
+      image: 'border-indigo-500',
+      outer: 'hover:border-indigo-500',
+    },
+    text: 'hover:text-indigo-600',
+  },
+  description: 'Guess where every NFL team is located on a map',
+  image: '/assets/project_images/NFLGeoguessr/NFLGeoguessr.jpg',
+  tags: ['Sports', 'Geography'],
+  link: '/nfl-geoguessr',
+  inProgress: false,
+};